Mock object for testing caching issue
Replaced LdaModel with a simple array to see if caching issue still
occurs. It does not. The problem appears to be related to stopwords. I
should test separating the corpus from the stopwords and apply the
stopwords each time a topic model is created, ie add preprocessing as
an explicit step at that time, not when stopword list is changed.