Tested the output of the document topic matrix
Need some more comprehensive testing with another corpus. One property
of topic models is that they are based on word co-locations, so a topic
might be highly correlated with other documents associated with the
topic even though it does not contain any or most of the top-ten
keywords for this topic. Especially when the documents are short.